+++ Sources/Graphics/Graphics+Stroke.swift
@@ -8,14 +8,14 @@ public extension Graphics {
 
     struct Stroke : Equatable {
         
-        public let width: Float
+        public let width: Double
         public let join: Graphics.LineJoin
         public let cap: Graphics.LineCap
         public let dash: Graphics.LineDash?
         public let fill: Graphics.Fill
         
         public init(
-            width: Float,
+            width: Double,
             join: Graphics.LineJoin = .miter,
             cap: Graphics.LineCap = .butt,
             dash: Graphics.LineDash? = nil,
