+++ Sources/Graphics/Protocol/IGraphicsRuleGuide.swift
@@ -6,6 +6,6 @@ import Foundation
 
 public protocol IGraphicsRuleGuide : IGraphicsGuide {
     
-    func guide(_ value: Float) -> Float?
+    func guide(_ value: Double) -> Double?
 
 }
