+++ Sources/RemoteImage/Protocol/IRemoteImageTarget.swift
@@ -6,7 +6,7 @@ import Foundation
 
 public protocol IRemoteImageTarget : AnyObject {
     
-    func remoteImage(progress: Float)
+    func remoteImage(progress: Double)
     func remoteImage(image: UI.Image)
     func remoteImage(error: Error)
