+++ Sources/RemoteImage/RemoteImage+Cache+Config+Memory.swift
@@ -8,10 +8,10 @@ public extension RemoteImage.Cache.Config {
 
     struct Memory {
 
-        public let maxImageArea: Float
+        public let maxImageArea: Double
         
         public init(
-            maxImageArea: Float = 2048 * 2048
+            maxImageArea: Double = 2048 * 2048
         ) {
             self.maxImageArea = maxImageArea
         }
