+++ Sources/UI/Protocol/Container/IUIBookContainer.swift
@@ -9,9 +9,9 @@ public protocol IUIBookContainer : IUIContainer, IUIContainerParentable {
     var backward: IUIBookContentContainer? { get }
     var current: IUIBookContentContainer? { get }
     var forward: IUIBookContentContainer? { get }
-    var animationVelocity: Float { set get }
+    var animationVelocity: Double { set get }
 #if os(iOS)
-    var interactiveLimit: Float { set get }
+    var interactiveLimit: Double { set get }
 #endif
     
     func reload(backward: Bool, forward: Bool)
