+++ Sources/UI/Protocol/Container/IUIGroupContainer.swift
@@ -7,14 +7,14 @@ import Foundation
 public protocol IUIGroupContainer : IUIContainer, IUIContainerParentable {
     
     var bar: UI.View.GroupBar { get }
-    var barVisibility: Float { get }
+    var barVisibility: Double { get }
     var barHidden: Bool { get }
     var containers: [IUIGroupContentContainer] { get }
     var backward: IUIGroupContentContainer? { get }
     var current: IUIGroupContentContainer? { get }
     var forward: IUIGroupContentContainer? { get }
     var allowAnimationUserInteraction: Bool { set get }
-    var animationVelocity: Float { set get }
+    var animationVelocity: Double { set get }
     
     func updateBar(animated: Bool, completion: (() -> Void)?)
