+++ Sources/UI/Protocol/View/IUIViewAlphable.swift
@@ -6,14 +6,14 @@ import Foundation
 
 public protocol IUIViewAlphable : AnyObject {
     
-    var alpha: Float { set get }
+    var alpha: Double { set get }
     
 }
 
 public extension IUIViewAlphable where Self : IUIWidgetView, Body : IUIViewAlphable {
     
     @inlinable
-    var alpha: Float {
+    var alpha: Double {
         set { self.body.alpha = newValue }
         get { self.body.alpha }
     }
@@ -24,7 +24,7 @@ public extension IUIViewAlphable {
     
     @inlinable
     @discardableResult
-    func alpha(_ value: Float) -> Self {
+    func alpha(_ value: Double) -> Self {
         self.alpha = value
         return self
     }
