+++ Sources/UI/Protocol/View/IUIViewProgressable.swift
@@ -6,14 +6,14 @@ import Foundation
 
 public protocol IUIViewProgressable : AnyObject {
     
-    var progress: Float { set get }
+    var progress: Double { set get }
     
 }
 
 public extension IUIViewProgressable where Self : IUIWidgetView, Body : IUIViewProgressable {
     
     @inlinable
-    var progress: Float {
+    var progress: Double {
         set { self.body.progress = newValue }
         get { self.body.progress }
     }
@@ -24,7 +24,7 @@ public extension IUIViewProgressable {
     
     @inlinable
     @discardableResult
-    func progress(_ value: Float) -> Self {
+    func progress(_ value: Double) -> Self {
         self.progress = value
         return self
     }
