+++ Sources/UI/UI+Border.swift
@@ -9,7 +9,7 @@ public extension UI {
     enum Border : Equatable {
         
         case none
-        case manual(width: Float, color: UI.Color)
+        case manual(width: Double, color: UI.Color)
         
     }
